qgptp: Added fix for dead lock.

+++ daemons/gptp/common/ether_port.cpp
@@ -836,7 +836,9 @@ int EtherPort::getRxTimestamp
 void EtherPort::startPDelayIntervalTimer
 ( long long unsigned int waitTime )
 {
-	pDelayIntervalTimerLock->lock();
+	if (pDelayIntervalTimerLock->trylock() == oslock_fail) {
+		return;
+	}
 	clock->deleteEventTimerLocked(this, PDELAY_INTERVAL_TIMEOUT_EXPIRES);
 	clock->addEventTimerLocked(this, PDELAY_INTERVAL_TIMEOUT_EXPIRES, waitTime);
 	pDelayIntervalTimerLock->unlock();
